// src/app/shared/components/dynamic-form/classes/control-enums.ts
export enum ControlType {
  dropdown = 'dropdown',
  textbox = 'textbox',
  checkbox = 'checkbox',
  toggle = 'toggle',
  base = 'base',
  input = 'input',
  numberInput = 'numberInput',
  stringInput = 'stringInput',
  counterInput = 'counterInput',
  radio = 'radio',
  radioCard = 'radioCard',
  datepicker = 'datepicker',
  label = 'label',
  password = 'password',
  email = 'email',
}

export enum ControlWidthClass {
  W100 = 'w-100',
  W60 = 'w-60',
  W50 = 'w-49',
  W40 = 'w-40',
  W33 = 'w-33',
  W25 = 'w-25',
  W20 = 'w-20',
}

export enum CustomFormError {
  noError = '',
  mismatch = 'mismatch',
}

export enum ControlFullLabelWidthClass {
  'w-100' = 'w-100',
  'w-49' = 'w-200',
  'w-33' = 'w-300',
  'w-25' = 'w-400',
}

export enum ControlBaseInputTypes {
  password = 'password',
  number = 'number',
  text = 'text',
  tel = 'tel',
  email = 'email',
}

export enum CommonFormLangName {
  prefix = 'common.form',
  label = 'label',
  emptyLabel = 'emptyLabel',
  emptyPlaceholder = 'emptyPlaceholder',
}

export enum CheckboxLabelPlacement {
  start = 'start',
  end = 'end',
  fixed = 'fixed',
  stacked = 'stacked',
}

export enum CheckboxJustification {
  start = 'start',
  end = 'end',
  spaceBetween = 'space-between',
}

export enum CheckboxAlignment {
  start = 'start',
  center = 'center',
}

export enum ControlBackground {
  white = 'white-bg',
  grey = 'grey-bg',
  boxShadow = 'box-shadow-bg',
  transparent = 'transparent',
}

export enum CustomFormTemplate {

}

export enum ControlColor {
  primary = 'primary',
  secondary = 'secondary',
  tertiary = 'tertiary',
  success = 'success',
  warning = 'warning',
  danger = 'danger',
  light = 'light',
  medium = 'medium',
  dark = 'dark',
}